[Fix] AVX512: SIMD1_64f TRUNC/MTRUNC
Former-commit-id: f73a512d6411e4f24997a3bc2a879c85df86c39d

+++ plugins/avx512/float/UMESimdVecFloat64_1.h
@@ -1007,14 +1007,14 @@ namespace SIMD {
         }
         // MROUND
         // TRUNC
-        inline SIMDVec_i<int32_t, 1> trunc() const {
-            int32_t t0 = (int32_t)mVec;
-            return SIMDVec_i<int32_t, 1>(t0);
+        inline SIMDVec_i<int64_t, 1> trunc() const {
+            int64_t t0 = (int64_t)mVec;
+            return SIMDVec_i<int64_t, 1>(t0);
         }
         // MTRUNC
-        inline SIMDVec_i<int32_t, 1> trunc(SIMDVecMask<1> const & mask) const {
-            int32_t t0 = mask.mMask ? (int32_t)mVec : 0;
-            return SIMDVec_i<int32_t, 1>(t0);
+        inline SIMDVec_i<int64_t, 1> trunc(SIMDVecMask<1> const & mask) const {
+            int64_t t0 = mask.mMask ? (int64_t)mVec : 0;
+            return SIMDVec_i<int64_t, 1>(t0);
         }
         // FLOOR
         // MFLOOR
